// types/teamType.ts
export type TeamType = {
  champExperience: number;
  championName: string;
  kills: number;
  deaths: number;
  assists: number;
  item0: number;
  item1: number;
  item2: number;
  item3: number;
  item4: number;
  item5: number;
  item6: number;
  totalMinionsKilled: number;
  neutralMinionsKilled: number;
  wardsKilled: number;
  summonerName: string;
  puuid: string;
};